+++ packages/react-cosmos-voyager/src/fixture-extensions.js
